add more awkward particule testing for #737
