+++ scratch.js
@@ -20,9 +20,18 @@ const nlp = require('./src/index')
 // let doc = nlp(`i am nice, he really drive`)
 // let doc = nlp(`john will drive the car!`)
 
-let doc = nlp(`Japan 'Twitter killer' pleads guilty to murders`)
+// let doc = nlp('i am being driven')
+let doc = nlp('i should be driven')
+// let doc = nlp('i should drive')
+// let doc = nlp('i should have been driven')
+// doc.verbs().toPastTense()
+doc.verbs().toFutureTense()
+doc.debug()
+// doc.verbs().toFutureTense().debug()
+
+// let doc = nlp(`Japan 'Twitter killer' pleads guilty to murders`)
 // let doc = nlp(`pleads`)
-doc.nouns().debug()
+// doc.nouns().debug()
 // let m = doc.verbs().subject().debug()
 // console.log(m)
 // m.lookBehind().nouns(null, { keep_anaphora: true }).last().debug()
